EEH-2513 - Merged the rest of functions/classes in utils
